add pessimistic lock

+++ api-server/src/main/java/me/blog/count/DailyCountRepository.java
@@ -9,5 +9,5 @@ import org.springframework.data.jpa.repository.Lock;
 public interface DailyCountRepository extends JpaRepository<DailyCount, Long> {
 
     @Lock(value = LockModeType.PESSIMISTIC_WRITE)
-    Optional<DailyCount> findByArticleIdAndDate(long articleId, LocalDate date);
+    Optional<DailyCount> findByArticleIdAndDate(int articleId, LocalDate date);
 }
